pyproject.toml
[tool.poetry]
name = "network-monitoring"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
prometheus-client = "^0.10.1"
psycopg2-binary = "^2.8.6"
psycopg2 = "^2.8.6"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
